+++ docs/changelog.rst
@@ -37,6 +37,9 @@ fixes
   name when the file haven't extension in ideamFormat2jazikuFormat
 - Fix special character in stations name and name variable in
   continue station in ideamFormat2jazikuFormat script
+- Fix calculating the number of days in month with data daily and
+  fixed when the year is repeat for the same station in
+  ideamFormat2jazikuFormat script
 - Fix units in particular type of time series in y-label for graphs
   for inspection of series
 - Fixes units variable translations for var_D and var_I